requirements.txt
click
python-telegram-bot==20.6
retry
APScheduler==3.10.4
tzlocal==4.3
bs4
requests
httpx
XClientTransaction
playwright
aiohttp